% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkInvertedLatitudeLongitude}
\alias{checkInvertedLatitudeLongitude}
\title{Check If Longitude and Latitude Are Inverted}
\usage{
checkInvertedLatitudeLongitude(
  lat,
  lon,
  latzone = c(-90, 90),
  longzone = c(-180, 180)
)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{latzone}{range of valid latitudes}

\item{longzone}{range of valid longitudes}
}
\value{
index of inverted values
}
\description{
Check If Longitude and Latitude Are Inverted
}
