% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkLatitude}
\alias{checkLatitude}
\title{Check If Latitude is In The Right Interval}
\usage{
checkLatitude(lat, minmax = c(-90, 90))
}
\arguments{
\item{lat}{latitude in decimal format}

\item{minmax}{range of latitude}
}
\value{
the index of each value out of range.
}
\description{
Check If Latitude is In The Right Interval
}
