% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkLongitude}
\alias{checkLongitude}
\title{Check If Longitude is In The Right Interval}
\usage{
checkLongitude(lon, minmax = c(-180, 180))
}
\arguments{
\item{lon}{longitude in decimal format}

\item{minmax}{range of longitude}
}
\value{
the index of each value out of range.
}
\description{
Check If Longitude is In The Right Interval
}
