% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkMissing}
\alias{checkMissing}
\title{Check Missing Values}
\usage{
checkMissing(vec)
}
\arguments{
\item{vec}{vector of values (n>=1)}
}
\value{
index of missing values
}
\description{
Check Missing Values
}
