% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dateTimeConverter}
\alias{dateTimeConverter}
\title{Convert DateTime To Chron}
\usage{
dateTimeConverter(date, time)
}
\arguments{
\item{date}{date value}

\item{time}{time value}
}
\value{
chron datetime value
}
\description{
Convert DateTime To Chron
}
