% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fromChronToDate}
\alias{fromChronToDate}
\title{Covert Chron Numeric Date Format to Date Time}
\usage{
fromChronToDate(x)
}
\arguments{
\item{x}{numeric value of date}
}
\value{
Date Time
}
\description{
Covert Chron Numeric Date Format to Date Time
}
