% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_file_name}
\alias{get_file_name}
\title{Extract File Name}
\usage{
get_file_name(file)
}
\arguments{
\item{file}{complete path with name of file}
}
\value{
file name
}
\description{
Extract the file name from a path
}
