% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_file_path}
\alias{get_file_path}
\title{Extract File Directory}
\usage{
get_file_path(file)
}
\arguments{
\item{file}{complete path with name of file}
}
\value{
file path
}
\description{
Extract the file directory from a path
}
