% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lastIndexOf}
\alias{lastIndexOf}
\title{Last Index Of}
\usage{
lastIndexOf(string, character)
}
\arguments{
\item{string}{a string to be checked}

\item{character}{a character that we look for}
}
\value{
index
}
\description{
returns the last index of a character in a string.
}
