% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maia_create_db.R
\name{maia_create_db}
\alias{maia_create_db}
\title{Create Relational VMS Data Base}
\usage{
maia_create_db(
  file,
  fao_species_code = "./fao_codes/FAO_CODES_SPECIES.csv",
  add = FALSE
)
}
\arguments{
\item{file}{the .vms file}

\item{fao_species_code}{FAO codes for species}

\item{add}{if TRUE append to existent data, FALSE (default)}
}
\value{
Doesn't return a value
}
\description{
creates an SQLite database for VMS data. It creates also additional tables
that could be used for other purposes of analyzing and cleaning the data.
}
