% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maia_edit_raw.R
\name{maia_edit_raw}
\alias{maia_edit_raw}
\title{Format MAIA csv File into Rmaiavms Formmat}
\usage{
maia_edit_raw(file, vms_db_path)
}
\arguments{
\item{file}{path to the landings data}

\item{vms_db_path}{path to the vms database}
}
\value{
this function doesn't return a value
}
\description{
Format MAIA csv File into Rmaiavms Formmat
}
