% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maia_vms_join.R
\name{maia_vms_join}
\alias{maia_vms_join}
\title{Joining VMS and MAIA data}
\usage{
maia_vms_join(logbook_db_path, vms_db_path)
}
\arguments{
\item{logbook_db_path}{MAIA database path}

\item{vms_db_path}{VMS database path}
}
\value{
no return
}
\description{
Joining VMS and MAIA data
}
