% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_assign_bathy.R
\name{vms_assign_bathy}
\alias{vms_assign_bathy}
\title{Assign Depth to VMS Pings}
\usage{
vms_assign_bathy(vms_db_path, use_res = 2, use_alg = "Slow & Light")
}
\arguments{
\item{vms_db_path}{path the SQLite database of VMS data}

\item{use_res}{Resolution of data}

\item{use_alg}{Algorithm to use "Slow & Light" (default) or "Fast & Heavy"}
}
\value{
this function doesn't return a value
}
\description{
Assign Depth to VMS Pings
}
