% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_assign_fishing.R
\name{vms_assign_fishing}
\alias{vms_assign_fishing}
\title{Title}
\usage{
vms_assign_fishing(
  vms_db_path,
  harb_file_path,
  min_vel_km = 3.7,
  max_vel_km = 11.11,
  min_depth_mt = -5000,
  max_depth_mt = -5,
  harb_dist_km = 16
)
}
\arguments{
\item{vms_db_path}{vms databse path}

\item{harb_file_path}{harbor shapefile}

\item{min_vel_km}{minimum velocity of fishing}

\item{max_vel_km}{maximum velocity of fishing}

\item{min_depth_mt}{min depth of fishing}

\item{max_depth_mt}{max depth of fishing}

\item{harb_dist_km}{harbor distance}
}
\value{

}
\description{
Title
}
