% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_clean_db.R
\name{vms_clean_db}
\alias{vms_clean_db}
\title{Clean the VMS data in the SQLite database}
\usage{
vms_clean_db(vms_db_path, map_file_path, harb_file_path, inHarbBuff = 2)
}
\arguments{
\item{vms_db_path}{path to the SQLite database}

\item{map_file_path}{shapefile of the land}

\item{harb_file_path}{shapefile of the positions of harbours}

\item{inHarbBuff}{distance from harbour that considered in harbour}
}
\value{
Edit the database warn table
}
\description{
Clean the VMS data in the SQLite database
}
