% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_create_db.R
\name{vms_create_db}
\alias{vms_create_db}
\title{Create RSQLite VMS Data Base}
\usage{
vms_create_db(vms_file, add = FALSE)
}
\arguments{
\item{vms_file}{the .vms file}

\item{add}{if TRUE append to existent data, FALSE (default)}
}
\value{
Doesn't return a value
}
\description{
creates an SQLite database for VMS data. It creates also additional tables
that could be used for other purposes of analyzing and cleaning the data.
}
