% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_cut_tracks.R
\name{vms_cut_tracks}
\alias{vms_cut_tracks}
\title{Cut VMS Pings into Tracks}
\usage{
vms_cut_tracks(
  vms_db_path,
  use_bp_out = TRUE,
  use_med_out = TRUE,
  use_ud_freq = 120
)
}
\arguments{
\item{vms_db_path}{VMS SQLite database path}

\item{use_bp_out}{parameter for outlier manipulation}

\item{use_med_out}{parameter for outlier manipulation}

\item{use_ud_freq}{parameter for outlier manipulation}
}
\value{
This function doesn't have a return value
}
\description{
Cut VMS Pings into Tracks
}
