% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_edit_raw.R
\name{vms_edit_raw}
\alias{vms_edit_raw}
\title{Format TACSAT VMS csv File into Rmaiavms Formmat}
\usage{
vms_edit_raw(
  vms_tacsat_file,
  latzone = c(-90, 90),
  longzone = c(-180, 180),
  append = FALSE,
  file_to_append = ""
)
}
\arguments{
\item{vms_tacsat_file}{csv file of VMS data in TACSAT format}

\item{latzone}{latitude range of the studied zone}

\item{longzone}{longitude range of the studied zone}

\item{append}{logical, if TRUE append old file with the new data}

\item{file_to_append}{file to append}
}
\value{
VMS file in rmaiavms format
}
\description{
Format TACSAT VMS csv File into Rmaiavms Formmat
}
