% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms_interpolate_tracks.R
\name{vms_interpolate_tracks}
\alias{vms_interpolate_tracks}
\title{Interpolates VMS tracks into low resolution}
\usage{
vms_interpolate_tracks(vms_db_path, int_fre = 10)
}
\arguments{
\item{vms_db_path}{path to VMS SQLite database}

\item{int_fre}{interpolation frequency in minutes}
}
\value{
This function doesn't have a return value
}
\description{
Interpolates VMS tracks into low resolution
}
