% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmv-package.R
\docType{package}
\name{rmv-package}
\alias{rmv}
\alias{rmv-package}
\title{rmv: What the Package Does (One Line, Title Case)}
\description{
This package allows the interpolation of VMS ping data, the joint of VMS positions and declared catch and the identification of fishing operations.
}
\author{
\strong{Maintainer}: Abdelouahed Ben Mhamed \email{a.benmhamed@intelligentica.net} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
